Polygon meshing(splitting into grids) methodology
In order to mesh the polygon into a 5x5 grid sections, it was required to first mask the polygon over a finite reference region.
We utilized poly2mask with a region size of 2*Radius x 2*Radius. The masked polygon from MaskPolygon.m was meshed into 25 grids (5x5) in MeshPolygon.m which outputs a cell that contains the polygon's masked values as a matrix. ReconstructPolygon.m can be used to approximately reform the polygon outline.